+++ lib/eventlog/regress/eventlog_store/store_sudo_test.c
@@ -51,7 +51,7 @@ compare(FILE *fp, const char *infile, const char *output)
 
     /* Expect two log lines, one for accept, one for exit. */
     for (i = 0; i < 2; i++) {
-	ssize_t output_len = strcspn(output, "\n");
+	ssize_t output_len = (ssize_t)strcspn(output, "\n");
 	ssize_t len = getdelim(&line, &linesize, '\n', fp);
 	if (len == -1) {
 	    sudo_warn("getdelim");
@@ -60,7 +60,7 @@ compare(FILE *fp, const char *infile, const char *output)
 	if (line[len - 1] == '\n')
 	    len--;
 
-	if (len != output_len || strncmp(line, output, len) != 0) {
+	if (len != output_len || strncmp(line, output, (size_t)len) != 0) {
 	    fprintf(stderr, "%s: %s mismatch\n", infile, i ? "exit" : "accept");
 	    fprintf(stderr, "expected: %.*s\n", (int)len, line);
 	    fprintf(stderr, "got     : %.*s\n", (int)output_len, output);
